Add tests where they belong
