+++ tests/value/test_stopping.py
@@ -8,9 +8,11 @@ from pydvl.utils import Status
 from pydvl.value import ValuationResult
 from pydvl.value.stopping import (
     HistoryDeviation,
+    MaxChecks,
     MaxTime,
     MaxUpdates,
     MinUpdates,
+    StandardError,
     StoppingCriterion,
     make_criterion,
 )
@@ -153,3 +155,44 @@ def test_history_deviation(n_steps, rtol):
         status |= done(v)
 
     assert status == Status.Converged
+
+
+def test_standard_error():
+    """Test the StandardError stopping criterion."""
+    eps = 0.1
+    n = 5
+
+    done = StandardError(threshold=eps)
+
+    # Trivial case: no variance.
+    v = ValuationResult(values=np.ones(n), variances=np.zeros(n))
+    assert done(v)
+
+    # Reduce the variance until the criterion is triggered.
+    v = ValuationResult(values=np.ones(n), variances=np.ones(n))
+    assert not done(v)
+
+    # One value is being left out
+    for _ in range(10):
+        for idx in range(1, n):
+            v.update(idx, 1)
+    assert not done(v)
+
+    # Update the final value
+    for _ in range(10):
+        v.update(0, 1)
+    assert done(v)
+
+
+def test_max_checks():
+    """Test the MaxChecks stopping criterion."""
+    v = ValuationResult.from_random(size=5)
+
+    done = MaxChecks(None)
+    for _ in range(10):
+        assert not done(v)
+
+    done = MaxChecks(5)
+    for _ in range(4):
+        assert not done(v)
+    assert done(v)
